Fix permission for mitmproxy directory
Fix issue https://github.com/mitmproxy/docker-releases/issues/6.

This approach is the same used by official Docker images such as mysql and postgres.

Running the container with any mitmproxy tool it will set the required permissions for
mitmproxy directory and run the process also with the mitmproxy user.

Otherwise it will run the passed arg with the root user. Which allows the user to
test the container with any command. Just will not set the directory permission running this way.