Change setup logic from filename to recursive folder search
Originally, we had a single folder of processed logs, with each log given a
distinct filename.  We used this filename to populate our api and fill out the
release.  This wasn't easy to scale however, for the amount of files we're
adding.  And so we switched to looking for any .json's inside our
data-gen/processed folder, which will be filled with an ever-growing series of
folders and files.  This required us to change our backend setup config to do a
recursive search, and to generate the release from the metadata file instead of
the filename.