Master release based on latest job in gci-gce bucket
Before we used to look for a file called master.  Now, we are working with
dynamic folders, and master is just whatever bit of info is in the latest job
inside the gci-gce bucket.  To be able to show this correctly required a
somewhat heavy rewrite of our releases bundle in the client, since it was built
around this notion of 'master'.  We've done this, but had to disable some of the
ui so the site wouldn't break.  This should be a  fairly straightforward fix
next though.