alter bucketjobrelease to not include dev/
Originally our audit logs were stored in processed/dev/buckets/job.  They are
now stored in processed/buckets/job.  We pull out the bucket and job using a
basic string split, and grabbing the right indices in the new array.  This
adjustment updates the indices to account for the lack of dev/