Have backend server  ignore changes to client files
We have a nodemon script set for the backend server, that triggers a restart
whenever a file has changed in the backend directory.  Since the backend dir now
includes our client files, and the client files change a lot during development,
this was causing a bunch of unnecessary restarts.  We now have the backend
nodemon ignore all file changes within client.  IT will restart when we change
the backend, and nothing else