useful filter query params more explicit
Before, when clicking on a link it would retain any existing query param and
then append the links filters onto it.  This causes issues when someone has set
a useragent filter, then switches to one of our links that does not have a
useragent explicitly defined.  this causes the previus useragent filte to carry
over, which does not match the intent of our filter link.  Now clicking on a
link will reset all filters except ones the link explicitly sets, and the bucket
we are currently looking at.