api_operations has no select subqueries nor requires raw_paths
We had an interim view for our operations called raw_paths, but I was able to
clean up the logic for our api_operations to no longer need this.  In addition,
we had 2 or 3 select sub-queries, which can greatly hamper performance.  Now,
everything is done through lateral joins with no nested subqueries