move older tables org to archive, rename t_and_v_bot.org
This is to provide clarity when navigating the repo and knowing where to add
code.  I didn't want to delete tables_and_views.org as it still has some
interesting tables and experiments in it, but they aren't a part of apisnoop at
the moment.  the bot version has become the default version of snoop, and so I
renamed it accordingly