'New Coverage' in percentage graph only shows tech debt
this is a revert from prior commit.  If we mark new coverage as 'new endpoints
with tests' and 'old endpoints with tests' it makes for a confusing graph, where
it can look liked ocverage goes down.  Now, we are only distinguishing 'old
endpoints with tests' to show when coverage has gone up through writing tests.
The new endpoints added with tests are just included in our total coverage