previousRelease predicate checks for null value
JS throws an error if you are trying to access a property of a null value.
Our NewEndpoints value is based on the endpoints that existed in the
previousRelease, and so we check those endpoints as part of this
function.

Our earliest release, by definition, has no previous release and so by
checking for its endpoints, we trigger this error.

This updated predicate fixes this issue, by checking for a values
existence before trying to access any part of it.