Check for TESTING_ONLY env in snoopdb startup
This passes in a TESTING_ONLY env if it is found in the instance creation.
With this set, no audit events are loaded in snoopdb's startup. This is
so someone can start up snoopdb and load events manually, to debug this
process, or so someone can use a slim test-writing only version of
snoopdb, since test writers don't necessarily need the full coverage
details.

The aim was to make this env var the lightest intrusion into how things
currently run on snoopdb. Helm charts can initialize env vars in the
snoopdb container, env vars that are then accessible in plpython. This
means we can keep everything as is, but if the env var is present, have
the loading events function return early instead of going through the
full process.