use artifacts/metadata.json to get job version
This swaps out the file we use to determine which version of kubernetes
was used in the test run.  Previously we used /finished.json, but there
are more cases coming up where a successful test run has a new type of
finished.json that does not include the kubernetes version.  Both our
older job runs and the newer ones includes /artifacts/metadata.json,
though, and it's 'job-version' key contains the info we need.

Signed-off-by: Zach Mandeville <zz@ii.coop>