ignore $ substitution error
this function will throw an error on the regex ending with a dollar sign,
as this is plpython doing string templating and interpreeting that
dollar sign as a variable invocation.  We can ignore this though and do
the substitution as is, treating the dollar sign as a dollar sign, using safe_substitute.

Ultimately, we should look at a way to run this without using pl/python
and directly as sql.  It will make it less finicky.

Signed-off-by: zim <webmaster@coolguy.website>