add delete_older_test_events function
This brings in a helper function that takes an interval( like '2 hours'
or '10 minutes' ) and deletes all events in testing.audit_event older than
that interval. It is meant to help  test writers who are generating a
lot of events and want to keep their event table fast and relevant.

Signed-off-by: zim <webmaster@coolguy.website>