Remove --kube-root deprecated kind argument
The --kube-root flag has been deprecated in kind for some time and will
be removed. The current equivalent functionality is to pass the path to
the kubernetes source as the first non-flag argument to "build
node-image".

This updates the kind instructions to use the newer preferred method in
preparation for kind dropping the flag.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>